refactor(orchestrator): modify code to use new paths
