+++ orchestrator/cmd/main.go
@@ -25,7 +25,7 @@ import (
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
-	"github.com/openclarity/vmclarity/pkg/orchestrator"
+	orchestrator "github.com/openclarity/vmclarity/orchestrator/pkg"
 	"github.com/openclarity/vmclarity/utils/log"
 	"github.com/openclarity/vmclarity/utils/version"
 )
