+++ orchestrator/pkg/assetscanprocessor/processor.go
@@ -25,7 +25,7 @@ import (
 	"github.com/openclarity/vmclarity/api/client"
 	"github.com/openclarity/vmclarity/api/types"
 	"github.com/openclarity/vmclarity/cli/pkg/utils"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/common"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/common"
 	"github.com/openclarity/vmclarity/utils/log"
 )
