+++ orchestrator/pkg/orchestrator.go
@@ -24,20 +24,20 @@ import (
 
 	"github.com/openclarity/vmclarity/api/client"
 	"github.com/openclarity/vmclarity/api/types"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/assetscanestimationwatcher"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/assetscanprocessor"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/assetscanwatcher"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/discovery"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/aws"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/azure"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/docker"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/external"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/gcp"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/kubernetes"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/scanconfigwatcher"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/scanestimationwatcher"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/scanwatcher"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/assetscanestimationwatcher"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/assetscanprocessor"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/assetscanwatcher"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/discovery"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/scanconfigwatcher"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/scanestimationwatcher"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/scanwatcher"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
+	"github.com/openclarity/vmclarity/provider/pkg/aws"
+	"github.com/openclarity/vmclarity/provider/pkg/azure"
+	"github.com/openclarity/vmclarity/provider/pkg/docker"
+	"github.com/openclarity/vmclarity/provider/pkg/external"
+	"github.com/openclarity/vmclarity/provider/pkg/gcp"
+	"github.com/openclarity/vmclarity/provider/pkg/kubernetes"
 	"github.com/openclarity/vmclarity/utils/log"
 )
