+++ orchestrator/pkg/scanwatcher/config.go
@@ -19,7 +19,7 @@ import (
 	"time"
 
 	"github.com/openclarity/vmclarity/api/client"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
 )
 
 const (
