+++ orchestrator/pkg/scanwatcher/watcher.go
@@ -25,8 +25,8 @@ import (
 	"github.com/openclarity/vmclarity/api/client"
 	"github.com/openclarity/vmclarity/api/types"
 	"github.com/openclarity/vmclarity/cli/pkg/utils"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/common"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
+	"github.com/openclarity/vmclarity/orchestrator/pkg/common"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
 	"github.com/openclarity/vmclarity/utils/log"
 )
