+++ provider/pkg/aws/provider.go
@@ -31,9 +31,9 @@ import (
 
 	"github.com/openclarity/vmclarity/api/types"
 	"github.com/openclarity/vmclarity/cli/pkg/utils"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/aws/scanestimation"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/cloudinit"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
+	"github.com/openclarity/vmclarity/provider/pkg/aws/scanestimation"
+	"github.com/openclarity/vmclarity/provider/pkg/cloudinit"
 	"github.com/openclarity/vmclarity/utils/log"
 )
