+++ provider/pkg/azure/common.go
@@ -23,7 +23,7 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
 )
 
 func handleAzureRequestError(err error, actionTmpl string, parts ...interface{}) (bool, error) {
