+++ provider/pkg/azure/scannerVm.go
@@ -26,8 +26,8 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/network/armnetwork/v3"
 
 	"github.com/openclarity/vmclarity/cli/pkg/utils"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/cloudinit"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
+	"github.com/openclarity/vmclarity/provider/pkg/cloudinit"
 )
 
 var (
