+++ provider/pkg/external/helpers.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/openclarity/vmclarity/api/types"
 	"github.com/openclarity/vmclarity/cli/pkg/utils"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
-	provider_service "github.com/openclarity/vmclarity/pkg/orchestrator/provider/external/proto"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
+	provider_service "github.com/openclarity/vmclarity/provider/pkg/external/proto"
 )
 
 func convertAssetToModels(asset *provider_service.Asset) (types.Asset, error) {
