+++ provider/pkg/external/provider.go
@@ -24,8 +24,8 @@ import (
 	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/openclarity/vmclarity/api/types"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
-	provider_service "github.com/openclarity/vmclarity/pkg/orchestrator/provider/external/proto"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
+	provider_service "github.com/openclarity/vmclarity/provider/pkg/external/proto"
 )
 
 type Provider struct {
