+++ provider/pkg/gcp/scannerVm.go
@@ -23,8 +23,8 @@ import (
 	"cloud.google.com/go/compute/apiv1/computepb"
 
 	"github.com/openclarity/vmclarity/cli/pkg/utils"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider/cloudinit"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
+	"github.com/openclarity/vmclarity/provider/pkg/cloudinit"
 )
 
 var (
