+++ provider/pkg/kubernetes/provider.go
@@ -35,7 +35,7 @@ import (
 	"github.com/openclarity/vmclarity/cli/pkg/families"
 	"github.com/openclarity/vmclarity/cli/pkg/utils"
 	discoveryclient "github.com/openclarity/vmclarity/containerruntimediscovery/client"
-	"github.com/openclarity/vmclarity/pkg/orchestrator/provider"
+	provider "github.com/openclarity/vmclarity/provider/pkg"
 )
 
 type Provider struct {
